+++ src/config.php
@@ -34,6 +34,9 @@ return [
     // API key for the Imgix transform service
     'imgixApiKey' => '',
 
+    // The optional security token used to sign image URLs from Imgix
+    '$imgixSecurityToken' => '',
+
     // Should image variant be created on Asset save (aka BeforePageLoad)
     'generateTransformsBeforePageLoad' => true,